cookbook/pydanticai/requirements.txt

python-dotenv
openai
pydantic
pydantic-ai
requests